pyproject.toml
[build-system]
requires = ["setuptools>=67.0"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-voice-transcriber"
version = "0.1.0"
description = "Export and transcribe Telegram messages into Markdown."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "telethon>=1.34.0",
    "faster-whisper>=1.0.0",
    "tqdm>=4.66.0",
    "python-dateutil>=2.8.2",
    "tzlocal>=5.2",
    "rich>=13.7.0",
    "typer>=0.12.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.5",
    "freezegun>=1.5.1",
]

[project.scripts]
tg-transcribe = "telegram_voice_transcriber.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
